modularize to npm
