+++ npm/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "files": ["main.ts"],
+    "compilerOptions": {
+        "strict": true,
+        "noImplicitAny": true,
+        "removeComments": true,
+        "declaration": true,
+        "checkJs": true,
+        "target": "ES2015",
+        "module": "CommonJS",
+        "lib": ["ES2019"],
+        "outDir": "dist"
+    }
+}
