+++ test-utils.js
@@ -1,6 +1,5 @@
 // @ts-check
 
-const { diffStringsUnified } = require('jest-diff');
 /** @param {string[]} lines */
 const join = (...lines) => lines.join('\n');
 
@@ -312,21 +311,6 @@ Cats have 9 lives`,
         </div>
     </div>`,
     },
-};
-
-/**
- *
- * @param {import("tape").Test} t
- * @param {string} actual
- * @param {string} expected
- * @param {string} [message]
- */
-module.exports.testString = function (t, actual, expected, message) {
-    t.equals(actual, expected, message);
-
-    if (actual !== expected) {
-        console.log(diffStringsUnified(expected, actual));
-    }
 };
 
 module.exports.possibleArguments = [
