+++ vscode/package.json
