+++ vscode/src/comment-regexs.ts
