+++ vscode/test/parse-string-arguments.test.js
